Missing cast

+++ tools/checksum_bitflip/checksum_bitflip.c
@@ -80,7 +80,7 @@ main(int argc, char *argv[])
     buffer = dmalloc(size);
 
     io_bytes = read(fd, buffer, size);
-    testutil_assert(io_bytes == size);
+    testutil_assert(io_bytes == (ssize_t)size);
 
     /* See if the checksum matches the file contents. */
     if (__wt_checksum_sw(buffer, size) == cksum_target) {
